+++ tools/samtools/pkg/samtools/view_posix_live_oracle_test.go
@@ -45,7 +45,7 @@ type decodedRec struct {
 	QName string
 	Flag  uint16
 	RName string
-	Pos   int32
+	Pos   int64
 	MapQ  uint8
 	Cigar string
 	Seq   string
